added price -1 for medicines that don't have valid pricelist

+++ isabackend/src/main/java/rs/ac/uns/ftn/isa/onee2team/isabackend/repository/IMedicineRepository.java
@@ -12,6 +12,6 @@ public interface IMedicineRepository extends JpaRepository<Medicine, Long> {
 	public List<Medicine> findMedicineByPharmacyid(Long id);
 
 	@Query("select w.medicine.id from Warehouse w where w.pharmacy.id = ?1")
-	public List<Integer> findMedicineIdsByPharmacyid(Long id);
+	public List<Long> findMedicineIdsByPharmacyid(Long id);
 
 }
